fix: fix ui_command_queue empty in release mode with -fvisibility=hidden flags.
Former-commit-id: 7105ef8dab16e0871245de49ab466c2602ce7477 [formerly f1be53bc06023d99c47ccb378c74e5335dfe45a4]
Former-commit-id: dc299acf11a175828bbfdd73066f009a363981e1

+++ bridge/include/kraken_bridge.h
@@ -106,7 +106,7 @@ KrakenInfo *getKrakenInfo();
 KRAKEN_EXPORT
 UICommandItem **getUICommandItems(int32_t contextId);
 KRAKEN_EXPORT
-size_t getUICommandItemSize(int32_t contextId);
+int64_t getUICommandItemSize(int32_t contextId);
 KRAKEN_EXPORT
 void clearUICommandItems(int32_t contextId);
